+++ src/cli.js
@@ -3,6 +3,7 @@ import { hookCommand, activateCommand, deactivateCommand } from "./commands/hook
 import { statusCommand } from "./commands/status.js";
 import { listCommand } from "./commands/list.js";
 import { editCommand } from "./commands/edit.js";
+import { linkCommand } from "./commands/link.js";
 import { doctorCommand } from "./commands/doctor.js";
 import { versionCommand } from "./commands/version.js";
 import { helpCommand } from "./commands/help.js";
@@ -36,6 +37,9 @@ export async function run(argv = process.argv) {
       break;
     case "edit":
       await editCommand();
+      break;
+    case "link":
+      await linkCommand(rest);
       break;
     case "doctor":
       await doctorCommand();